refactor: internal/nodecfg
