+++ internal/db/multiaccounts/settings/database.go
@@ -13,8 +13,8 @@ import (
 	"github.com/status-im/status-go/crypto/types"
 	maErrors "github.com/status-im/status-go/internal/db/multiaccounts/errors"
 	"github.com/status-im/status-go/internal/db/sqlite"
+	"github.com/status-im/status-go/internal/nodecfg"
 	"github.com/status-im/status-go/logutils"
-	"github.com/status-im/status-go/nodecfg"
 	"github.com/status-im/status-go/params"
 )
