+++ pkg/backend/geth_backend.go
@@ -43,9 +43,9 @@ import (
 	"github.com/status-im/status-go/internal/images"
 	"github.com/status-im/status-go/internal/instrumentation/trace"
 	"github.com/status-im/status-go/internal/metrics"
+	"github.com/status-im/status-go/internal/nodecfg"
 	"github.com/status-im/status-go/internal/transactions"
 	"github.com/status-im/status-go/logutils"
-	"github.com/status-im/status-go/nodecfg"
 	"github.com/status-im/status-go/params"
 	"github.com/status-im/status-go/pkg/backend/node"
 	nodeadapters "github.com/status-im/status-go/pkg/backend/node/adapters"
